api-gateway-bedrock-ip-restriction-waf-codepipeline/requirements.txt
aws-cdk-lib==2.185.0
python-dotenv==1.1.0
constructs>=10.0.0,<11.0.0